+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: Added filter by latest state
 
 1.2.0 (2022-04-14)
 ==================